% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conclusions.R
\name{ci_conclusion}
\alias{ci_conclusion}
\title{Descriptive conclusions from confidence intervals.}
\usage{
ci_conclusion(ci, groups = c("Control intervention", "Test intervention"),
  beneficial_outcome = TRUE, test_type = c("superiority", "noninferiority",
  "equivalence"), margin = 0, actual_null = 0)
}
\arguments{
\item{ci}{A single row from a matrix of the type returned by confint(),
containing the confidence interval for the parameter estimate. The two
columns provide the lower and upper confidence limits.}

\item{groups}{Character vector of length 2 containing descriptive labels for the
the interventions received by each group, with the control group label
first and the intervention group label second.}

\item{beneficial_outcome}{Is the outcome to be treated as beneficial
(i.e., a higher value of the outcome is superior)?}

\item{test_type}{Type of test to be conducted (superiority, noninferiority or equivalence).}

\item{margin}{Numerical value specifying the non-inferiority or equivalence
margin to be used if conducting those test types.}

\item{actual_null}{The value that precisely zero difference would have in
the parameter being examined. For an absolute measure this will typically
be 0. For a relative measure it will typically be 1. For superiority tests
this is the point value that the confidence interval is compared at. For
non-inferiority / equivalence tests it is the starting point that the
margin is applied to in order to establish the point / region for
comparison.}
}
\value{
A list object with elements stating the conclusion in different
  formats.
}
\description{
Produces a descriptive conclusion from a confidence interval, depending
on the type of test specified (superiority, non-inferiority or equivalence).
}
\examples{
# Establish a test confidence interval
ci_test <- matrix(c(-0.1,0.1),
                  nrow = 1, dimnames = list("estimate",
                                            c("2.5 \%","97.5 \%")))
ci_conclusion(ci_test)
ci_conclusion(ci_test, test_type = "noninferiority", margin=0.15)
ci_conclusion(ci_test, test_type = "equivalence", margin=0.15)
ci_conclusion(ci_test, test_type = "equivalence", margin=0.05)

}
