% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence_difference.R
\name{prevalence_difference}
\alias{prevalence_difference}
\title{Prevalence difference of an outcome between groups.}
\usage{
prevalence_difference(outcome_data, groups, outcomes)
}
\arguments{
\item{outcome_data}{Data providing the outcomes against the groups. Can be provided as raw
(i.e., participant-level) data, detailing row-by-row the the group and
outcome for each participant. Can also be provided as aggregate data.
Valid input types are:
  data.frame with 2 columns and a row for each participant, where the
    first column contains the participant's group, second column
    contains the participant's outcome;
  matrix containing a contingency table of the groups against the
    outcomes, with 2 columns (one for the 'outcome occurred' state and
    one for the 'outcome did not occur' state), and a named row for each
    group;
  list with 4 named elements, group1_t, group1_f, group2_t and group2_f,
    giving the number of participants in each category, where group1 is
    the control group and group2 is the intervention group, and t is the
    'outcome occurred' state and f is the 'outcome did not occur' state.}

\item{groups}{Character vector of the groups in the order in which they should be
analysed, starting with the control group. For list data the parameter
will define the labels to be applied in the output, and default values
will be used if not specified.}

\item{outcomes}{Character vector of length 2 specifying the labels used in the data for
the two outcome states, with the 'outcome occurred' state listed first and
the 'outcome did not occur' state second. For list data the parameter will
define the labels to the be applied in the output, and default values
will be used if not specified.}
}
\value{
The prevalence difference, indicating the extent to which the outcome
  is more prevalent in group 2 than group 1. Returns a list of objects if
  there are more than 2 groups.
}
\description{
Calculates the prevalence difference of a binary / dichotomous outcome
between two or more groups, from individual or aggregate data.
}
\details{
The prevalence difference is defined for binary / dichotomous outcomes in a
trial. In each group, the prevalence is the number of participants for whom
the outcome occured as a proportion of the total number of participants in
that group. For example, if 15 participants out of 100 in a group achieved
the outcome in question, the prevalence for that group would be 0.15 (15%).

The prevalence difference is the difference between the prevalence of the
outcome in an intervention group and the prevalence in the control group.

As an absolute measure, the prevalence difference tends to create a
relatively accurate impression of effect size. Some guides recommend
presenting relative measures as well as absolute measures. If looking for
a comparable relative measure, consider \code{\link{prevalence_ratio}}.

The function is able to handle both individual and aggregate data.
Individual data should be provided as a data.frame with two columns.
Aggregate data can be provided as either a named list or a matrix with
named rows and columns.
}
\examples{
dummy_rct_data_list   <- list(group1_t = 48, group1_f = 52,
                              group2_t = 64, group2_f = 36)
prevalence_difference(outcome_data = dummy_rct_data_list,
                      groups = c("Control group", "Intervention group"),
                      outcomes = c("Outcome occurred", "Outcome did not occur"))
# Data will be reordered if necessary.
dummy_rct_data_matrix <- matrix(data=c(36, 64,
                                       52, 48),
                                  ncol=2, byrow=TRUE,
                                  dimnames = list(c("Group 1 - new treatment",
                                                    "Control group"),
                                                  c("No","Yes")))
prevalence_difference(outcome_data = dummy_rct_data_matrix,
                      groups = c("Control group", "Group 1 - new treatment"),
                      outcomes = c("Yes", "No"))
# Multiple-arm studies are supported (4 groups in this example):
dummy_rct_4 <- matrix(c(7,21,20,30,93,79,80,70),
                      nrow=4,
                      dimnames = list(c("Treatment as usual",
                                        "Intervention 1",
                                        "Intervention 2",
                                        "Intervention 3"), c("Yes","No")))
pd4 <- prevalence_difference(dummy_rct_4,
                             groups = rownames(dummy_rct_4),
                             outcomes = colnames(dummy_rct_4))
# Returns a list of prevalence_difference objects.
length(pd4)  # Length 3, comparing each of interventions 1-3 to TAU.
pd4[[1]]

}
